Add multiplyMatrices method

+++ src/main.cpp
@@ -7,14 +7,14 @@
 int main() {
 
 	double termVec[constants::n];
-	double *coefficientMat[constants::n], *lowerTMat[constants::n], 
-	*upperTMat[constants::n], *diagonalMat[constants::n];
+	double *coefficientMat[constants::n], *lowerTMat[constants::n],
+		 *upperTMat[constants::n], *diagonalMat[constants::n];
 
 	for (auto i = 0; i < constants::n; i++) {
-		coefficientMat[i] = new double[constants::n]{0.0};
-		lowerTMat[i] = new double[constants::n]{0.0};
-		upperTMat[i] = new double[constants::n]{0.0};
-		diagonalMat[i] = new double[constants::n]{0.0};
+		coefficientMat[i] = new double[constants::n]{ 0.0 };
+		lowerTMat[i] = new double[constants::n]{ 0.0 };
+		upperTMat[i] = new double[constants::n]{ 0.0 };
+		diagonalMat[i] = new double[constants::n]{ 0.0 };
 	}
 
 	auto start = std::chrono::system_clock::now();
@@ -47,7 +47,7 @@ int main() {
 				break;
 			}
 		}
-		termVec[i] = sin((i+1) * (constants::f + 1));
+		termVec[i] = sin(((double)i + 1.0) * (constants::f + 1));
 	}
 
 	auto end = std::chrono::system_clock::now();
@@ -58,7 +58,23 @@ int main() {
 	// test_matrix(upperTMat, "Upper");	
 	// test_matrix(lowerTMat, "Lower");
 	// test_matrix(diagonalMat, "Diag");
-
+	//double *aMatrix[constants::nTest],  *bMatrix[constants::nTest];
+	//std::srand((size_t)time(NULL));
+	//for (auto i = 0; i < constants::nTest; i++)
+	//{
+	//	aMatrix[i] = new double[constants::nTest];
+	//	bMatrix[i] = new double[constants::nTest];
+	//	for (auto j = 0; j < constants::nTest; j++) {
+	//		aMatrix[i][j] = (double)(std::rand() % 10);
+	//		bMatrix[i][j] = (double)(std::rand() % 10);
+	//	}
+	//}
+	//test_matrix(aMatrix, "A");
+	//test_matrix(bMatrix, "B");
+	//double **cMatrix = addMatrices(aMatrix, bMatrix);
+	//test_matrix(cMatrix, "Test sum");
+	//double **dMatrix = multiplyMatrices(aMatrix, bMatrix);
+	//test_matrix(dMatrix, "Test product");
 	system("pause");
 	return 0;
 }
